Add self-loops to edgelist and create edge features/labels
Add function to add self-loops to graph before generating the edgelist. Functions that create the heads/tails matrices should reference an edgelist to ensure a fixed ordering of edges.

Also created the modified feature and label set, which correspond to edges. The only labels of interest are those which correspond to self-loops, which represent the hidden state of each node.

Test plan: pytest